The HTML!
Probally the best one in the alpha, bcoz it's minimalist and working!
The install script is fine so you can install it!
Unfortunately, it doesn't work in vanillaOS bcoz you can't modify any directory like /var /etc or /bin.
Works great on NixOS.
Tested and developed on NixOS 23.11 Tapir.
Works also on arch based distros(and arch of course), and on debian and derivateds.
Please know if you have our unique and never used before dependencie, we presents, from the shadows, GIT! Install it please.