Update install.sh
This is version 0.0.2 alpha. 
The biggest improvement is the fact that it now has the tmp folder which is basically what stores the app data during installation, but then it is empty, because it is useless to leave the application remnants itself, and also (within the biggest improvement) there is the repos folder which is where you can put the app repositories! The install script is (without a doubt) the feature that will be updated the least. LOL